Add BloomCausalLM (#1467)
* Initial commit for BloomCausalLm

* Avoid adding a start token into token ids

* Revert "Avoid adding a start token into token ids"

This reverts commit 57ce4c203b40bdfc8c8361987df25349029cecfc.

* Tie embeddding weights

* Export BloomCausalLM

* Add bloomz to the preset map

* Fix some presets names

* Revert "Fix some presets names"

This reverts commit 1b5949ffae58fef20b2a931638a2813230b4f96a.

* Add doc Example

* format the code

* Alibi bias small fixes

* Add tests

* Maek float16 dtype test to keras3 only

* Update model version

* Edit exampples

* Remove max_sequnce_length argument

* Try fix keras2 error

* Update hf_model download fun

* Make 1b models easier to copy

* Optimize conversion script

* Save checkpoints in float16

* Add test for mixed_float16

* Try to reproduce the keras2 Error

* Revert "Try to reproduce the keras2 Error"

This reverts commit 7cc8671097277c3c2503c0fe6d98facabe87ecf6.

* Revert "Make 1b models easier to copy"

This reverts commit 4d701d3331dd1c8db079b2e1f1fd2a974a72f244.

* Show How to couple dtype_policy between backbone, causalLM, and backbone layers in dtype arg is based to bacckbone

* Revert "Show How to couple dtype_policy between backbone, causalLM, and backbone layers in dtype arg is based to bacckbone"

This reverts commit 77c576bf2002d4c3889bcc83444b870edcad3e7e.

* Add validate_only flag to conversion script

* Change preset version

* set cache dtype to self.compute_dtype

* Minor fix

+++ keras_nlp/layers/modeling/alibi_bias.py
@@ -94,7 +94,9 @@ class AlibiBias(keras.layers.Layer):
         )
         slopes = ops.expand_dims(slopes, 1)
 
-        seq_range = ops.expand_dims(ops.arange(1 - key_length, 1), 0)
+        seq_range = ops.expand_dims(
+            ops.arange(1 - key_length, 1, dtype="int32"), 0
+        )
         seq_range = ops.cast(seq_range, dtype=self.compute_dtype)
 
         alibi_bias = ops.multiply(slopes, seq_range)
